infra/_framework-pkg/_framework/_fw_repos_diagram_exporter/requirements.txt
pyyaml>=6.0
ruamel.yaml>=0.17
